apps/inference/requirements.txt
flask
flask-cors
scikit-learn
joblib
pandas
numpy
pillow
opencv-python
statsmodels
tensorflow
